requirements.txt
Flask==1.1.1
mongoengine==0.19.1
mongomock==3.18.0
flask-mongoengine==0.9.5
py-kim==1.2.4
flasgger==0.9.4
pytest==5.3.5
PyHamcrest==2.0.0
